"hide" KDNode in details too
